move shared types out of next-app-loader/index
